Redis Learning

+++ BootWithDatabase/src/main/java/com/zensar/repository/CompanyRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.zensar.entity.Company;
 
 
-public interface CompanyRepo extends CrudRepository<Company, Long> {
+public interface CompanyRepo extends CrudRepository<Company, Integer> {
 }
