+++ BootWithDatabase/src/main/java/com/zensar/repository/EmployeeRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 import com.zensar.entity.Employees;
 
 
-public interface EmployeeRepo extends CrudRepository<Employees, Long> {
+public interface EmployeeRepo extends CrudRepository<Employees, Integer> {
 
 
 }
