+++ Kubernetes_with_DB/src/main/java/com/k8s/repository/EmployeeRepo.java
@@ -3,7 +3,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.k8s.entity.EmployeesData;
 
-public interface EmployeeRepo extends CrudRepository<EmployeesData, Long> {
+public interface EmployeeRepo extends CrudRepository<EmployeesData, Integer> {
 
 
 }
